food-nutrition-backend/requirements.txt
flask
flask-cors
torch
torchvision
Pillow